Reorganize react components, add initial styles
