+++ public/index.html
@@ -13,9 +13,9 @@
     <link rel="stylesheet" href="/bundle.css">
 </head>
 <body>
-    <div id="app">
+    <main id="app" class="app">
         <%- initialHtml %>
-    </div>
+    </main>
 
     <script src="/bundle.js"></script>
 </body>
