--- src/frontend/scripts/components/Login.jsx
@@ -1,11 +1,0 @@
-var React = require('react');
-
-
-var Login = React.createClass({
-
-    render: function render() {
-        return <h1>Login</h1>;
-    }
-});
-
-module.exports = Login;
