--- src/frontend/scripts/components/NotFound.jsx
@@ -1,9 +1,0 @@
-var React = require('react');
-
-var NotFound = React.createClass({
-    render: function render() {
-        return <h1>NotFound</h1>;
-    }
-});
-
-module.exports = NotFound;
