--- src/frontend/scripts/components/Project.jsx
@@ -1,10 +1,0 @@
-var React = require('react');
-
-
-var Project = React.createClass({
-    render: function render() {
-        return <h1>Project</h1>;
-    }
-});
-
-module.exports = Project;
