+++ src/frontend/scripts/pages/NotFoundPage.jsx
@@ -1,0 +1,14 @@
+var React = require('react');
+
+
+var NotFoundPage = React.createClass({
+    render: function render() {
+        return (
+            <div className="page">
+                <h2>Page not found</h2>
+            </div>
+        );
+    }
+});
+
+module.exports = NotFoundPage;
