+++ src/frontend/scripts/pages/PostPage.jsx
@@ -1,0 +1,15 @@
+var React = require('react');
+
+
+var Project = React.createClass({
+    render: function render() {
+        return (
+            <div className="page">
+                <h1>Project page</h1>
+                <h3>Mozilla</h3>
+            </div>
+        );
+    }
+});
+
+module.exports = Project;
