+++ src/frontend/styles/general.styl
@@ -1,3 +1,6 @@
-body, html
+body, html, .app, .app > *
   width 100%
   height 100%
+  margin 0
+  padding 0
+  display block
