+++ src/frontend/styles/layout.styl
@@ -1,0 +1,12 @@
+.app > *
+  display flex
+
+.page
+  flex 1
+
+.menu-bar
+  position absolute
+  right 0
+  background #eee
+  height 100%
+  width 200px
