+++ src/frontend/styles/main.styl
@@ -2,4 +2,5 @@
 
 @import 'variables'
 @import 'fonts'
+@import 'layout'
 @import 'general'
